requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-filter>=24.0
django-environ>=0.11.0
drf-spectacular>=0.27.0
pytest>=8.0
pytest-django>=4.8.0
requests>=2.28.0
rich>=13.0.0
